+++ CHANGES.rst
@@ -283,6 +283,10 @@ API changes
   - WCS objects can now be initialized with an ImageHDU or
     PrimaryHDU object. [#4493]
 
+  - astropy.wcs now issues an INFO message when the header has SIP coefficients but
+    "-SIP" is missing from CTYPE. [#4814]
+
+
 Bug fixes
 ^^^^^^^^^
 
@@ -378,6 +382,11 @@ Bug fixes
     failing as a side-effect of #4699. [#4757]
 
 - ``astropy.wcs``
+
+  - astropy.wcs.to_header removes "-SIP" from CTYPE when SIP coefficients
+    are not written out, i.e. ``relax`` is either ``False`` or ``None``.
+    astropy.wcs.to_header appends "-SIP" to CTYPE when SIP coefficients
+    are written out, i.e. ``relax=True``. [#4814]
 
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^